requirements.txt
feedparser==6.0.11
httpx>=0.27.0
beautifulsoup4==4.12.3
playwright==1.49.0
openai>=1.0.0
python-telegram-bot==21.3
python-dotenv==1.0.1
pystray>=0.19.0
Pillow>=10.0.0
customtkinter>=5.2.0
schedule>=1.2.0
pytest==8.2.0
pytest-mock==3.14.0
